Fix Test Suite
This commit fixes the test suite, which was horribly broken because of
the new version of truffle / web3 / ethereumjs being used to accommodate
solidity version 0.7. There were several breaking API changes which
needed to be fixed across the test suite. This commit does not add any
new tests for the new functionality yet - that will be added in a future
change

CLOSES TICKET: BG-23871