Update sequenceID to be strictly increasing
The sequenceID logic before had a window that allowed for not strictly
increasing sequenceIDs, as long as they're within a certain window and
no single id is used more than once. This overly complicated logic
increases the gas usage of transfers, and isn't really necessary, as we
already have to strictly order transactions due to the nonces for the
sending address. This commit removes that logic and makes it only
strictly increasing