Add linter, fix lints and static analysis
This commit adds the solhint linter with the recommended settings, and
a few tweaks that are required by our use case (for example, sending
using .call). It fixes all of the lints. I also ran the slither static
analysis tool and fixed the issues there that made sense to fix.