Move batcher functionality into the wallet contract
This commit removes the separate Batcher contract and creates a new
function on walletsimple called sendMultiSigBatch. It takes a list of
recipients and values, and a user signature over those, plus the usual
expiration date and sequenceId. It then loops through and sends funds to
each recipient for the specified value to create a batched transaction.
This was moved into the main contract to remove the issues associated
with getting funds stuck on an external contract due to failure, which
allowed us to simplify the batching functionality significantly.