Update coin-specific wallet contracts
This commit updates the coin-specific wallets - ETC, RSK, and CELO.
These coins are basically the same as the ETH wallet but they have
different network ids that are signed over by the user signature to
avoid cross-chain replays of signatures. In order to avoid entirely
duplicating the code I added virtual functions to the main wallet so the
coin-specific wallets just have to override those functions and add
their own network ids.