Add documentation to CloneFactory
The CloneFactory is heavily inspired by the EIP-1167 minimal proxy reference
implementation. That implementation has assembly and bytecode to define
a proxy contract in a very small footprint. This commit adds explanatory
comments to describe to a reader of the code what the assembly and
bytecode are referring to in this section of the code. This is in
response to the CoinSpect audit, issue BTG4-002