Add events for token transactions
This commit adds events for sendMultiSigToken and flushTokens which were
previously not logged with events. This will enable easier debugging,
and integration with external services to understand what is going on
with Bitgo wallets. These were previously not included as for our
purposes, the token's Transfer event was enough to capture the balance
updates and transfers. This is an improvement based on the CoinSpect
audit, specifically issue BTG-004