Disallow 0 address as signer, check invalid signature
This commit updates the wallet contract to throw if any of the given
signers are the 0 address, and reverts if the given signature is
invalid. This avoids a vulnerability where, if a wallet had a 0 address
as a signer, then an attacking co-signer could submit any invalid
signature and it would be taken as valid. This is because the ecrecover
function returns 0 on an invalid signature, which would match the 0x0
address in the signers list. This is part of the CoinSpect audit,
specifically issue BTG-001