Remove batcher gas checks
This commit removes the gas checks from the batcher, and makes it so
that if any transfer fails, or if the batch runs out of gas, the entire
batch fails and all funds are safely returned to the user. This is
a simpler model to conceptualize and integrate with, though will require
careful gas estimation and if any recipients fail on receive, could be
frustrating for users as they will have to retry the entire batch.

Ticket: BG-25667