Fix ForwarderDeposited event data
The ForwarderDeposited event in the Forwarder.sol contract was always
emitting with the from address set to the forwarder itself. This is not
the desired functionality - it should be set to the actual sender that
sent in to the forwarder in the first place.

For background, DELEGATECALL preserves the correct msg.sender from the
original caller. I.e. if A (sender) CALL -> B (proxy) DELEGATECALL
-> C (implementation), then in C, the msg.sender will be A. However,
CALLs change the msg.sender to the most recent address in the callstack.
functions marked as `external` can only be called externally, they so
they always force a `CALL`.

In this case, the `flush()` function was external, the path looked like:
A (sender) -> B (proxy) -> DELEGATECALL C (implementation receive())
Then the receive() caused B (proxy) -> CALL B (flush()) -> DELEGATECALL
C (flush()). That CALL changes the msg.sender to B instead of A, so our
event emitted by B had itself as the msg.sender.

This commit fixes the issue by making flush() a public function, which
can be called both externally and internally. It also changes the
callsites to remove the `this.` syntax so the calls will happen
internally when desired. We also add a new test to ensure that this
functionality works and continues to work in the future

CLOSES TICKET: BG-28248