Remove sending back stuck funds in batcher
Currently, sending extra funds to batcher will send all funds that are
in the batcher contract to msg.sender. We replace this action by
letting an admin call the recover function to remove stuck funds.
This is done to prevent someone from front running a transaction
that leaves ETH in the contract and calling batch() right after to
withdraw the contract balance.

CLOSES Ticket: BG-28164