Fix WalletSimple sequenceID handling
The WalletSimple multisig contract validates a sequence ID with each
signature to protect against signature replays. In the previous version
of the contract, this sequenceId was validated against a moving window
of the last 10 ids. This allowed for some flexibility with transaction
ordering, while avoiding the same id being used twice, and avoiding
keeping much state around. In the new version, we deemed this
unnecessary and simplified it to a purely increasing value. However,
it turns out that the ordering flexilibity is actually desirable and
worth the cost. This commit brings back the original moving window
implementation.