Read sequence id array from memory
Sload is a very expensive operation in ethereum - it costs 800 gas each
time. In the tryInsertSequenceId function we perform many sloads to
fetch the values of sequence ids at different locations in the array.
This commit improves the gas utilization of this function by only
fetching the values from storage one time, reading them from memory
after that.

CLOSES TICKET: BG-30046