feat: pack sequence ids to save gas
The sequence id system for WalletSimple is quite well bounded - the
array of recently used ids has a max length of 10 and their values
generally stay quite low. This commit changes the storage model for the
array to use a single uint256 with 10 tightly packed uint24 values. This
saves 9 sloads over the course of a sendMultiSig, saving about 10k gas.