WIP/POC: Clone Immutable Args
this commit is a proof of concept for using immutable args with clones
for the deployed WalletSimple and Forwarder clones. Immutable args are
much cheaper to read than storage args, and using the library
clones-with-immutable-args we can get effectively immutable cost on
proxy/clones. We had to tweak the library a bit to get it to work
properly with create2, and some tests are still failing.

This saves about 55k gas for deploying wallets and 4k for sending
transfers